chore: move prefixes in package binary to their own file
